chore: refactor to ee namespace
