+++ app/models/application_record.rb
@@ -5,11 +5,13 @@ class ApplicationRecord < ActiveRecord::Base
   before_validation :validates_column_content_length
 
   # the models that exposed in email templates through liquid
-  DROPPABLES = %w[Account Channel Conversation Inbox User Message SlaPolicy].freeze
+  def droppables
+    %w[Account Channel Conversation Inbox User Message]
+  end
 
   # ModelDrop class should exist in app/drops
   def to_drop
-    return unless DROPPABLES.include?(self.class.name)
+    return unless droppables.include?(self.class.name)
 
     "#{self.class.name}Drop".constantize.new(self)
   end
@@ -47,3 +49,5 @@ class ApplicationRecord < ActiveRecord::Base
     end
   end
 end
+
+ApplicationRecord.include_mod_with('Enterprise::ApplicationRecord')
