+++ enterprise/app/models/application_record.rb
@@ -1,0 +1,5 @@
+module Enterprise::ApplicationRecord
+  def droppables
+    super + %w[SlaPolicy]
+  end
+end
