+++ spec/mailers/agent_notifications/conversation_notifications_mailer_spec.rb
@@ -105,43 +105,4 @@ RSpec.describe AgentNotifications::ConversationNotificationsMailer do
       expect(mail).to be_nil
     end
   end
-
-  describe 'sla_missed_first_response' do
-    let(:sla_policy) { create(:sla_policy, account: account) }
-    let(:mail) { described_class.with(account: account).sla_missed_first_response(conversation, agent, sla_policy).deliver_now }
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq("Conversation [ID - #{conversation.display_id}] missed SLA for first response")
-    end
-
-    it 'renders the receiver email' do
-      expect(mail.to).to eq([agent.email])
-    end
-  end
-
-  describe 'sla_missed_next_response' do
-    let(:sla_policy) { create(:sla_policy, account: account) }
-    let(:mail) { described_class.with(account: account).sla_missed_next_response(conversation, agent, sla_policy).deliver_now }
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq("Conversation [ID - #{conversation.display_id}] missed SLA for next response")
-    end
-
-    it 'renders the receiver email' do
-      expect(mail.to).to eq([agent.email])
-    end
-  end
-
-  describe 'sla_missed_resolution' do
-    let(:sla_policy) { create(:sla_policy, account: account) }
-    let(:mail) { described_class.with(account: account).sla_missed_resolution(conversation, agent, sla_policy).deliver_now }
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq("Conversation [ID - #{conversation.display_id}] missed SLA for resolution time")
-    end
-
-    it 'renders the receiver email' do
-      expect(mail.to).to eq([agent.email])
-    end
-  end
 end
